Fix logic error
If TRIMPATH isn't set, then $tf will not be defined on line 423 and the
resulting call to fpart will fail. On line 311, we start the process of
handling the TRIMPATH but if we do that, then $tf becomes the "correct"
source list, not $FROMLIST. On line 376, make sure that $FROMLIST is now
set to the tempfile $tf which contains the trimmed paths we want to
sync.